app/requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.35.0
websockets==12.0
opencv-python==4.11.0.86
torch==2.8.0
torchvision==0.23.0
mediapipe==0.10.21
numpy
pandas==2.2.3
pillow==11.0.0
python-multipart==0.0.18
httpx==0.28.1
python-dotenv==1.0.1
pydantic==2.11.0
ultralytics
tqdm
seaborn
requests